Extract custom functions from storage entity block
