+++ framework/src/modules/chain/components/storage/entities/index.js
@@ -16,6 +16,7 @@
 
 module.exports = {
 	Account: require('./account'),
+	Block: require('./block'),
 	Migration: require('./migration'),
 	Peer: require('./peer'),
 	Round: require('./round'),
