+++ framework/src/modules/chain/components/storage/sql/blocks/delete.sql
